# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/addUser": {
            "post": {
                "summary": "Adds a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/uint64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorApiMessage"
                        }
                    }
                }
            }
        },
        "/designs/add": {
            "post": {
                "summary": "Adds a design",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of current device",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "time of connection: 2012-11-01T22:08:41+00:00",
                        "name": "time",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "device connected to",
                        "name": "other",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "power of the connection",
                        "name": "power",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "rssi of the connection",
                        "name": "rssi",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/uint64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorApiMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "web.ErrorApiMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}